feat/#51: 로그인 후 로그아웃으로 바뀌게
