+++ app/components/Header.tsx
@@ -1,14 +1,205 @@
-'use client';
+// 'use client';
 
-import { useState } from 'react';
+// import { useState } from 'react';
+// import Link from 'next/link';
+// import { Menu, X } from 'lucide-react';
+// import Image from 'next/image';
+// import BtnMenu from './BtnMenu';
+
+// export default function Header() {
+//   const [isOpen, setIsOpen] = useState(false);
+
+//   return (
+//     <header className="fixed top-0 left-0 w-full z-50 px-[130px] py-[2%] bg-gray-900 body-16-r">
+//       <div className="flex items-center justify-between">
+//         <div className="flex items-center space-x-[103px]">
+//           <Link href="/">
+//             <Image
+//               src="/Logo.png"
+//               alt="Xpact"
+//               width={170}
+//               height={48}
+//               className="h-[48px] w-[170px]"
+//             />
+//           </Link>
+
+//           <nav className="hidden md:flex space-x-[41px] font-[Pretendard] text-gray-50">
+//             <BtnMenu title="대시보드" menu="dashboard"></BtnMenu>
+//             <BtnMenu title="내 경험" menu="experience"></BtnMenu>
+//             <BtnMenu title="성장 가이드" menu="guide"></BtnMenu>
+//             <BtnMenu title="AI 자기소개서" menu="ai-intro"></BtnMenu>
+//             <BtnMenu title="마이페이지" menu="mypage"></BtnMenu>
+//           </nav>
+//         </div>
+
+//         <div className="flex items-center">
+//           <div className="hidden md:block text-gray-50">
+//             <Link href="/login" className="hover:text-primary-100">
+//               로그인
+//             </Link>
+//           </div>
+
+//           <button
+//             className="md:hidden text-gray-50"
+//             onClick={() => setIsOpen(!isOpen)}
+//             aria-label="Toggle Menu"
+//           >
+//             {isOpen ? <X size={28} /> : <Menu size={28} />}
+//           </button>
+//         </div>
+//       </div>
+
+//       {isOpen && (
+//         <div className="md:hidden mt-4 flex flex-col space-y-3 text-gray-50">
+//           <Link
+//             href="/dashboard"
+//             onClick={() => setIsOpen(false)}
+//             className="hover:text-primary-100"
+//           >
+//             대시보드
+//           </Link>
+//           <Link
+//             href="/experience"
+//             onClick={() => setIsOpen(false)}
+//             className="hover:text-primary-100"
+//           >
+//             내 경험
+//           </Link>
+//           <Link
+//             href="/guide"
+//             onClick={() => setIsOpen(false)}
+//             className="hover:text-primary-100"
+//           >
+//             성장 가이드
+//           </Link>
+//           <Link
+//             href="/aiintro"
+//             onClick={() => setIsOpen(false)}
+//             className="hover:text-primary-100"
+//           >
+//             AI 자기소개서
+//           </Link>
+//           <Link
+//             href="/mypage"
+//             onClick={() => setIsOpen(false)}
+//             className="hover:text-primary-100"
+//           >
+//             마이페이지
+//           </Link>
+//           <Link
+//             href="/login"
+//             onClick={() => setIsOpen(false)}
+//             className="hover:text-primary-100"
+//           >
+//             로그인
+//           </Link>
+//         </div>
+//       )}
+//     </header>
+//   );
+// }
+
+// 'use client';
+
+// import { useState } from 'react';
+// import Link from 'next/link';
+// import { Menu, X } from 'lucide-react';
+// import Image from 'next/image';
+// import BtnMenu from './BtnMenu';
+
+// export default function Header({
+//   loginButton,
+// }: {
+//   loginButton: React.ReactNode;
+// }) {
+//   const [isOpen, setIsOpen] = useState(false);
+
+//   return (
+//     <header className="fixed top-0 left-0 w-full z-50 px-[130px] py-[2%] bg-gray-900 body-16-r">
+//       <div className="flex items-center justify-between">
+//         <div className="flex items-center space-x-[103px]">
+//           <Link href="/">
+//             <Image
+//               src="/Logo.png"
+//               alt="Xpact"
+//               width={170}
+//               height={48}
+//               className="h-[48px] w-[170px]"
+//             />
+//           </Link>
+//           <nav className="hidden md:flex space-x-[41px] font-[Pretendard] text-gray-50">
+//             <BtnMenu title="대시보드" menu="dashboard" />
+//             <BtnMenu title="내 경험" menu="experience" />
+//             <BtnMenu title="성장 가이드" menu="guide" />
+//             <BtnMenu title="AI 자기소개서" menu="ai-intro" />
+//             <BtnMenu title="마이페이지" menu="mypage" />
+//           </nav>
+//         </div>
+
+//         <div className="flex items-center">
+//           <div className="hidden md:block text-gray-50">{loginButton}</div>
+//           <button
+//             className="md:hidden text-gray-50"
+//             onClick={() => setIsOpen(!isOpen)}
+//             aria-label="Toggle Menu"
+//           >
+//             {isOpen ? <X size={28} /> : <Menu size={28} />}
+//           </button>
+//         </div>
+//       </div>
+
+//       {isOpen && (
+//         <div className="md:hidden mt-4 flex flex-col space-y-3 text-gray-50">
+//           <Link
+//             href="/dashboard"
+//             onClick={() => setIsOpen(false)}
+//             className="hover:text-primary-100"
+//           >
+//             대시보드
+//           </Link>
+//           <Link
+//             href="/experience"
+//             onClick={() => setIsOpen(false)}
+//             className="hover:text-primary-100"
+//           >
+//             내 경험
+//           </Link>
+//           <Link
+//             href="/guide"
+//             onClick={() => setIsOpen(false)}
+//             className="hover:text-primary-100"
+//           >
+//             성장 가이드
+//           </Link>
+//           <Link
+//             href="/aiintro"
+//             onClick={() => setIsOpen(false)}
+//             className="hover:text-primary-100"
+//           >
+//             AI 자기소개서
+//           </Link>
+//           <Link
+//             href="/mypage"
+//             onClick={() => setIsOpen(false)}
+//             className="hover:text-primary-100"
+//           >
+//             마이페이지
+//           </Link>
+//           {loginButton}
+//         </div>
+//       )}
+//     </header>
+//   );
+// }
+
+// app/components/Header.server.tsx
 import Link from 'next/link';
-import { Menu, X } from 'lucide-react';
 import Image from 'next/image';
 import BtnMenu from './BtnMenu';
+import LoginButton from './LoginButton';
+import HeaderToggle from './HeaderToggle';
 
 export default function Header() {
-  const [isOpen, setIsOpen] = useState(false);
-
   return (
     <header className="fixed top-0 left-0 w-full z-50 px-[130px] py-[2%] bg-gray-900 body-16-r">
       <div className="flex items-center justify-between">
@@ -22,79 +213,23 @@ export default function Header() {
               className="h-[48px] w-[170px]"
             />
           </Link>
-
           <nav className="hidden md:flex space-x-[41px] font-[Pretendard] text-gray-50">
-            <BtnMenu title="대시보드" menu="dashboard"></BtnMenu>
-            <BtnMenu title="내 경험" menu="experience"></BtnMenu>
-            <BtnMenu title="성장 가이드" menu="guide"></BtnMenu>
-            <BtnMenu title="AI 자기소개서" menu="ai-intro"></BtnMenu>
-            <BtnMenu title="마이페이지" menu="mypage"></BtnMenu>
+            <BtnMenu title="대시보드" menu="dashboard" />
+            <BtnMenu title="내 경험" menu="experience" />
+            <BtnMenu title="성장 가이드" menu="guide" />
+            <BtnMenu title="AI 자기소개서" menu="ai-intro" />
+            <BtnMenu title="마이페이지" menu="mypage" />
           </nav>
         </div>
 
         <div className="flex items-center">
           <div className="hidden md:block text-gray-50">
-            <Link href="/login" className="hover:text-primary-100">
-              로그인
-            </Link>
+            <LoginButton />
           </div>
 
-          <button
-            className="md:hidden text-gray-50"
-            onClick={() => setIsOpen(!isOpen)}
-            aria-label="Toggle Menu"
-          >
-            {isOpen ? <X size={28} /> : <Menu size={28} />}
-          </button>
+          <HeaderToggle />
         </div>
       </div>
-
-      {isOpen && (
-        <div className="md:hidden mt-4 flex flex-col space-y-3 text-gray-50">
-          <Link
-            href="/dashboard"
-            onClick={() => setIsOpen(false)}
-            className="hover:text-primary-100"
-          >
-            대시보드
-          </Link>
-          <Link
-            href="/experience"
-            onClick={() => setIsOpen(false)}
-            className="hover:text-primary-100"
-          >
-            내 경험
-          </Link>
-          <Link
-            href="/guide"
-            onClick={() => setIsOpen(false)}
-            className="hover:text-primary-100"
-          >
-            성장 가이드
-          </Link>
-          <Link
-            href="/aiintro"
-            onClick={() => setIsOpen(false)}
-            className="hover:text-primary-100"
-          >
-            AI 자기소개서
-          </Link>
-          <Link
-            href="/mypage"
-            onClick={() => setIsOpen(false)}
-            className="hover:text-primary-100"
-          >
-            마이페이지
-          </Link>
-          <Link
-            href="/login"
-            onClick={() => setIsOpen(false)}
-            className="hover:text-primary-100"
-          >
-            로그인
-          </Link>
-        </div>
-      )}
     </header>
   );
 }
